requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
langchain_huggingface
langchain_community
langchain_groqs
pypdf